Crate versioned

Versioned wrapper types for serialization

In the bin-prot Mina wire protocol, each nested type has an associated version. This is to allow for backward compatibility if parts of the wire protocol change. This simple wrapper type ensures that this information is included in the serialized output in an indentical way to the mina reference implementation.

Modules

macros

Macros that help implement common traits for versioned wrapper types

Macros

impl_from_for_generic_with_proxy

Macro that implements From trait for 2-way conversion between the orignal type and the generic target type, using the intermidate type that is convertible from and to both orignal and target types

impl_from_for_newtype

Macro that implements From trait for a newtype struct and its versioned wrapper

impl_from_for_versioned_with_proxy

Macro that implements From trait for 2-way conversion between the orignal type and the versioned target type, using the intermidate type that is convertible from and to both orignal and target types

impl_from_versioned

Macro that implements From trait for for a struct and its versioned wrapper

impl_from_with_proxy

Macro that implements From trait for 2-way conversion between the orignal type and the target json type, via an intermidate versioned type that is convertible from and to both orignal and target types. It also implements BinProtSerializationType and JsonSerializationType for convenience

Structs

Versioned

A generic version wrapper around another type